Add `AutoconfigFetcher` interface
